@@ -1,6 +1,20 @@
 {
   "type": "module",
   "dependencies": {
+    "@types/bcrypt": "^5.0.2",
+    "@types/bcryptjs": "^2.4.6",
+    "@types/compression": "^1.7.5",
+    "@types/cookie-parser": "^1.4.7",
+    "@types/cors": "^2.8.17",
+    "@types/express": "^4.17.21",
+    "@types/express-session": "^1.18.0",
+    "@types/ioredis": "^5.0.0",
+    "@types/jsonwebtoken": "^9.0.6",
+    "@types/mongoose": "^5.11.97",
+    "@types/passport": "^1.0.16",
+    "@types/passport-google-oauth20": "^2.0.16",
+    "@types/redis": "^4.0.11",
+    "@types/supertest": "^6.0.2",
     "axios": "^1.6.8",
     "bcryptjs": "^2.4.3",
     "compression": "^1.7.4",
@@ -8,32 +22,53 @@
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
-    "redis": "^4.6.13",
+    "express-session": "^1.18.0",
     "jsonwebtoken": "^9.0.2",
     "mongoose": "^8.0.0",
-    "nodemon": "^3.0.1"
+    "nodemon": "^3.0.1",
+    "passport": "^0.7.0",
+    "passport-google-oauth20": "^2.0.0",
+    "redis": "^4.6.13"
+  },
+  "lint-staged": {
+    "**/*.{js,ts,jsx,tsx}": [
+      "npx prettier --write",
+      "eslint --fix"
+    ]
   },
   "scripts": {
     "test": "jest --watchAll  --detectOpenHandles --verbose --coverage ",
-    "start": "nodemon server.js",
+    "start": "node server.js",
+    "dev": "nodemon server.js",
     "check": "prettier --check .",
     "format": "prettier --write .",
     "vercel-build": "echo yay"
   },
   "jest": {
-    "globalTeardown": "./tests/teardown.js",
+    "globalTeardown": "./tests/teardown",
     "transform": {
       "^.+\\.js?$": "babel-jest"
     }
   },
   "devDependencies": {
     "@babel/preset-env": "^7.23.2",
+    "@eslint/js": "^9.9.0",
+    "@jest/globals": "^29.7.0",
+    "@stylistic/eslint-plugin-js": "^2.1.0",
+    "@types/eslint": "^9.6.0",
+    "@types/jest": "^29.5.12",
+    "@typescript-eslint/eslint-plugin": "^8.1.0",
+    "@typescript-eslint/parser": "^8.1.0",
     "babel-jest": "^29.7.0",
-    "eslint": "^8.57.0",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-import": "^2.29.1",
+    "eslint": "^9.9.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-babel": "^5.3.1",
+    "eslint-plugin-prettier": "^5.1.3",
+    "globals": "^15.3.0",
     "jest": "^29.7.0",
-    "prettier": "^3.1.0",
-    "supertest": "^6.3.3"
+    "lint-staged": "^15.2.2",
+    "prettier": "^3.2.5",
+    "supertest": "^6.3.3",
+    "typescript": "^5.5.4"
   }
 }
