@@ -1,6 +1,6 @@
 /** @type {import('tailwindcss').Config} */
-const colors = require('tailwindcss/colors');
-const defaultTheme = require('tailwindcss/defaultTheme');
+import colors from 'tailwindcss/colors';
+import defaultTheme from 'tailwindcss/defaultTheme';
 
 export default {
   content: ['./index.html', './src/**/*.{js,ts,jsx,tsx}'],
@@ -17,6 +17,10 @@ export default {
           title: colors.slate[50],
           description: colors.slate[400],
           info: colors.slate[500],
+          field: colors.slate[900],
+          button: colors.slate[700],
+          textInField: colors.slate[50],
+          textColor: colors.slate[50],
         },
         light: {
           DEFAULT: colors.slate[50],
