@@ -14,7 +14,7 @@ export default function BlogFeed() {
   const [loading, setLoading] = useState(true);
 
   useEffect(() => {
-    let categoryEndpoint =
+    const categoryEndpoint =
       selectedCategory === 'featured'
         ? '/api/posts/featured'
         : `/api/posts/categories/${selectedCategory}`;
@@ -45,7 +45,7 @@ export default function BlogFeed() {
   return (
     <div className="mx-auto my-6">
       <div className="-mx-4 flex flex-wrap">
-        <div className="w-full p-4 md:w-2/3">
+        <div className="w-full p-4 sm:w-2/3">
           <div className="-mb-1 cursor-text text-base tracking-wide text-slate-500 dark:text-dark-tertiary">
             What's hot?
           </div>
@@ -64,7 +64,7 @@ export default function BlogFeed() {
                   .map((post, index) => <FeaturedPostCard key={index} post={post} />)}
           </div>
         </div>
-        <div className="w-full p-4 md:w-1/3">
+        <div className="w-full p-4 sm:w-1/3">
           <div className="mb-6">
             <div className="-mb-1 cursor-text text-base tracking-wide text-light-tertiary dark:text-dark-tertiary">
               Discover by topic
@@ -72,9 +72,10 @@ export default function BlogFeed() {
             <h2 className="mb-2 cursor-text text-xl font-semibold dark:text-dark-primary">
               Categories
             </h2>
-            <div className="flex flex-wrap gap-3 dark:rounded-lg dark:bg-dark-card dark:p-3">
+            <div className="flex flex-wrap gap-3">
               {categories.map((category) => (
                 <button
+                  name="category"
                   key={category}
                   aria-label={category}
                   type="button"
