@@ -1,12 +1,12 @@
 function hero() {
   return (
     <div className="mb-8 flex max-w-3xl flex-1 flex-col justify-end text-slate-50">
-      <h1 className="cursor-text text-xl font-bold md:text-4xl">Journey Beyond Horizons</h1>
-      <p className="my-4 cursor-text text-base md:text-lg">
+      <h1 className="cursor-text text-xl font-bold sm:text-4xl">Journey Beyond Horizons</h1>
+      <p className="my-4 cursor-text text-base sm:text-lg">
         Dive into the world of travel with stories that transport you to far-off lands. Adventure
         awaits around every corner. It's time to explore the world!
       </p>
-      <div className="cursor-text text-sm font-semibold md:text-xl">Let's go!</div>
+      <div className="cursor-text text-sm font-semibold sm:text-xl">Let's go!</div>
     </div>
   );
 }
