@@ -1,18 +1,15 @@
 import { useLayoutEffect, useState } from 'react';
 import Sun from '@/assets/svg/sun.svg';
 import Moon from '@/assets/svg/moon.svg';
+import useThemeClass from '@/utils/theme-changer';
 function ThemeToggle() {
   const [isDarkTheme, setIsDarkTheme] = useState<boolean | null>(null);
   const toggleTheme = () => {
     setIsDarkTheme((prevTheme) => (prevTheme === null ? true : !prevTheme));
   };
   useLayoutEffect(() => {
-    const storedTheme = localStorage.getItem('theme');
-    const prefersDark = window.matchMedia?.('(prefers-color-scheme: dark)').matches;
-    setIsDarkTheme(storedTheme === 'dark' || (!storedTheme && prefersDark) || null);
-    if (storedTheme === 'dark' || (!storedTheme && prefersDark)) {
-      document.documentElement.classList.add('dark');
-    }
+    const storedTheme = useThemeClass();
+    setIsDarkTheme(storedTheme === 'dark');
   }, []);
 
   useLayoutEffect(() => {
@@ -25,6 +22,7 @@ function ThemeToggle() {
   return (
     <div>
       <button
+        name="theme"
         className={`${
           isDarkTheme ? 'bg-dark-theme-background' : 'bg-light-theme-background'
         } flex h-8 w-16 cursor-pointer items-center justify-start rounded-full px-1 py-1`}
@@ -37,7 +35,7 @@ function ThemeToggle() {
               : 'translate-x-0 bg-light-theme-foreground'
           } h-6 w-6 rounded-full bg-black px-1 py-1 duration-300`}
         >
-          <img src={`${isDarkTheme ? Moon : Sun}`} />
+          <img src={`${isDarkTheme ? Moon : Sun}`} alt="theme-toggler" />
         </div>
       </button>
     </div>
