@@ -1,28 +1,29 @@
-import axios from 'axios';
 import { useEffect, useState } from 'react';
 import BlogFeed from '@/components/blog-feed';
 import PostCard from '@/components/post-card';
 import Post from '@/types/post-type';
 import { PostCardSkeleton } from '@/components/skeletons/post-card-skeleton';
 import Header from '@/layouts/header-layout';
+import axiosInstance from '@/helpers/axios-instance';
 function HomePage() {
   const [posts, setPosts] = useState<Post[]>([]);
 
   useEffect(() => {
-    axios
-      .get(import.meta.env.VITE_API_PATH + '/api/posts')
-      .then((response) => {
-        setPosts(response.data);
-      })
-      .catch((error) => {
+    const fetchPosts = async () => {
+      try {
+        const res = await axiosInstance.get('/api/posts');
+        setPosts(res.data);
+      } catch (error) {
         console.error(error);
-      });
+      }
+    };
+    fetchPosts();
   }, []);
 
   return (
     <div className="w-full cursor-default bg-light dark:bg-dark">
       <Header />
-      <div className="mx-4 md:mx-8 lg:mx-16">
+      <div className="mx-4 sm:mx-8 lg:mx-16">
         <BlogFeed />
         <h1 className="cursor-text pb-4 text-xl font-semibold dark:text-dark-primary sm:pb-0">
           All Posts
